Begin builder pattern
I wanted to expand the worksheet generator to do more than just "Fill in the blanks" type work sheets.

However, with the current structure of the codes this is impossible (unless I copied and pasted the codes multiple times).

I'm not attempted to follow the SOLID principles along with utilizing the Builder pattern to tidy everything up before I extend the script to generate more than one type of worksheet.